Move monkey-patch to lib
